+++ CHANGELOG.md
@@ -12,6 +12,8 @@ All changes to the **Hearts of Modding** extension will be documented in this fi
 
 - **Better semantic highlighting in events.**
 
+- **Meta-scope resolution refactor.**
+
 ### 🩹 Fixed
 
 - **Duplicate detection in vanilla files:** If your workspace has the same path as `hoi4.gamePath` (so your workspace is your specified HOI4 install) the extension will no longer give false warnings about duplicates.